requirements.txt
appdirs==1.4.4
appnope==0.1.0
asgiref==3.5.2
attrs==19.3.0
backcall==0.1.0
backports.zoneinfo==0.2.1
black==19.10b0
bleach==3.1.5
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
configparser==5.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==1.0.0
Django==4.1.4
django-cors-headers==3.1.1
django-heroku==0.3.1
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
drf-jwt==1.13.3
drf-yasg==1.20.0
entrypoints==0.3
filelock==3.0.12
gevent==21.8.0
greenlet==1.1.2
gunicorn==20.1.0
idna==2.8
image==1.5.32
inflection==0.5.1
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itypes==1.2.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.4
packaging==20.8
pandocfilters==1.4.2
parso==0.7.0
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psycopg2==2.8.6
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2020.4
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
regex==2020.6.8
requests==2.22.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
spotipy==2.15.0
sqlparse==0.4.1
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.11
virtualenv==20.0.21
wcwidth==0.1.9
webencodings==0.5.1
whitenoise==6.2.0
widgetsnbextension==3.5.1
zope.event==4.5.0
zope.interface==5.4.0
